+++ spec/sinatra/param_validator/validator/form_spec.rb
@@ -8,11 +8,11 @@ describe Sinatra::ParamValidator::Validator::Form do
     mock_app do
       register Sinatra::ParamValidator
 
-      validator type: :form, identifier: :identifier do
+      validator :identifier do
         param :number, Integer, required: true
       end
 
-      post '/', validate: :identifier do
+      post '/', validate_form: :identifier do
         'OK'.to_json
       end
     end