+++ spec/sinatra/param_validator/validator/url_param_spec.rb
@@ -8,11 +8,11 @@ describe Sinatra::ParamValidator::Validator::UrlParam do
     mock_app do
       register Sinatra::ParamValidator
 
-      validator type: :url_param, identifier: :identifier do
+      validator :identifier do
         param :number, Integer, required: true
       end
 
-      post '/', validate: :identifier do
+      post '/', validate_url_param: :identifier do
         'OK'.to_json
       end
     end